pyproject.toml
[tool.poetry]
name = "spam_detector"
version = "0.1.0"
description = "e-mail spam detector"
authors = ["Maldu <contactme@gmail.com>"]
readme = "README.md"
package-mode = false 

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = "^2.2.3"
kagglehub = "^0.3.4"
nltk = "^3.9.1"
xgboost = "^2.1.3"
dvc = "^3.57.0"
flask = "^2.2.1"
imblearn = "^0.0"
scikit-learn = "^1.5.2"
numpy = ">=1.18.5,<2.0"  
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
textattack = "^0.3.10"
joblib = "^1.4.2"
textaugment = "^2.0.0"
nlpaug = "^1.1.11"
ipywidgets = "^8.1.5"
boto3 = "^1.35.87"
torch = "^2.5.1"
transformers = "^4.47.1"
sentencepiece = "^0.2.0"
colorama = "^0.4.6"
huggingface-hub = {extras = ["cli"], version = "^0.27.0"}
sentence-transformers = "^3.3.1"
sacremoses = "^0.1.1"
apache-airflow = "^2.10.4"
matplotlib = "^3.10.0"

[tool.poetry.group.dev.dependencies]
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
ipykernel = "^6.29.5"
emoji = "^2.14.0"
mlflow = "^2.18.0"
wordcloud = "^1.9.4"
tqdm = "^4.67.1"

gensim = "4.3.3"
psutil = "^6.1.0"
pynvml = "^11.5.3"
papermill = "^2.6.0"
nbconvert = "^7.16.4"
webencodings = "^0.5.1"
notebook = "^7.3.1"
ipynbname = "^2024.1.0.0"
diskcache = "^5.6.3"
textaugment = "^2.0.0"

googletrans-python = "^0.2.0"

[tool.poetry.group.env.dependencies]
textblob = ">=0.15.3,<0.19.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



